NIAD-459 HTTP 500 when spine responds with unexpected response
